requirements.txt
falcon
firebase-admin
groq
jobspy
falcon-cors
python-dotenv
nltk
scikit-learn
pandas